requirements.txt
opencv-python==4.8.1.78
dlib==19.24.2
numpy==1.24.3
scipy==1.11.1
pyserial==3.5
cmake==3.27.2 